+++ src/explore-education-statistics-common/src/modules/table-tool/components/TableToolWizard.tsx
@@ -46,7 +46,7 @@ interface Publication {
   slug: string;
 }
 
-interface TableToolState {
+export interface TableToolState {
   query: TableDataQuery | undefined;
   createdTable: FullTable | undefined;
   validInitialQuery: TableDataQuery | undefined;
