+++ src/explore-education-statistics-common/src/services/dataBlockService.ts
@@ -2,6 +2,7 @@ import { dataApi } from '@common/services/api';
 import { Dictionary } from '@common/types/util';
 import { Feature, Geometry } from 'geojson';
 import { TableDataQuery } from '@common/modules/full-table/services/tableBuilderService';
+import { Table, Chart, Summary } from '@common/services/publicationService';
 
 export enum GeographicLevel {
   Establishment = 'Establishment',
@@ -282,18 +283,24 @@ export interface DataBlockRerequest {
   boundaryLevel?: number;
 }
 
+export interface DataBlock {
+  id?: string;
+
+  heading?: string;
+  customFootnotes?: string;
+  name?: string;
+  source?: string;
+
+  dataBlockRequest: DataBlockRequest;
+  charts?: Chart[];
+  tables?: Table[];
+  summary?: Summary;
+}
+
 export type DataBlockRequest = TableDataQuery;
 
-export interface DataBlockResponse {
+export interface DataBlockResponse extends DataBlockData {
   metaData: DataBlockMetadata;
-
-  publicationId: string;
-  releaseId: string;
-  subjectId: number;
-  releaseDate: Date;
-  geographicLevel: GeographicLevel;
-
-  result: Result[];
 }
 
 const DataBlockService = {
